aggregator/customer.go
package aggregator

import (
	"ddd/entity"
	"ddd/valueobject"
	"errors"
	"github.com/google/uuid"
)

var (
	ErrInvalidPerson = errors.New("Customer must have a name!!!")
)

type Customer struct {
	person       *entity.Person
	products     []*entity.Item
	transactions []valueobject.Transaction
}

func NewCustomer(name string) (Customer, error) {
	if name == "" {
		return Customer{}, ErrInvalidPerson
	}

	p := &entity.Person{
		Name: name,
		ID:   uuid.New(),
	}

	return Customer{
		person:       p,
		products:     make([]*entity.Item, 0),
		transactions: make([]valueobject.Transaction, 0),
	}, nil
}

func (c *Customer) GetId() uuid.UUID {
	return c.person.ID
}

func (c *Customer) SetId(uuid uuid.UUID) {
	if c.person == nil {
		c.person = &entity.Person{}
	}
	c.person.ID = uuid
}

func (c *Customer) GetName() string {
	return c.person.Name
}

func (c *Customer) SetName(name string) {
	if c.person == nil {
		c.person = &entity.Person{}
	}
	c.person.Name = name
}
